Updated CMakeLists + size_t type use.

+++ include/gui/models/Models.h
@@ -19,7 +19,7 @@ class PlayerListModel : public QAbstractTableModel {
         QVariant data(const QModelIndex &index, int role = Qt::DisplayRole) const override;
         QVariant headerData(int section, Qt::Orientation orientation, int role = Qt::DisplayRole) const override;
         void setPagination(int start, int max);
-        int totalPlayers() const;
+        size_t totalPlayers() const;
         int filteredPlayerCount() const;  
 
     public slots:
