+++ src/models/PlayerRating.cpp
@@ -81,7 +81,7 @@ void PlayerRating::processMatchesParallel(const std::vector<Game>& games, const 
     });
 
     #pragma omp parallel for ordered
-    for (size_t i = 0; i < sortedGames.size(); ++i) {
+    for (int i = 0; i < sortedGames.size(); ++i) {
         int gameId = sortedGames[i].gameId;
 
         if (gameAppearances.find(gameId) != gameAppearances.end()) {
